do not try to create ${INSTALL_DIR}/${TARGET}/usr symlink if it already exists

+++ Makefile
@@ -234,7 +234,7 @@ binutils: binutils-${VERSION_BINUTILS}-${CPU}-cross.ok
 gcc-${VERSION_GCC}-${CPU}-cross-preliminary.ok: gcc-${TARGET}.ok
 	rm -rf $@ ${FOLDER_GCC}-${CPU}-cross-preliminary
 	mkdir -p ${FOLDER_GCC}-${CPU}-cross-preliminary
-	ln -sfv . ${INSTALL_DIR}/${TARGET}/usr
+	test -L ${INSTALL_DIR}/${TARGET}/usr || ln -sfv . ${INSTALL_DIR}/${TARGET}/usr
 	cd ${FOLDER_GCC}-${CPU}-cross-preliminary && \
 	export PATH=${INSTALL_DIR}/bin:$$PATH && \
 	../${FOLDER_GCC}/configure \
